brain_lab/requirements.txt
pypdf==5.1.0
rich==13.9.4
pydantic==2.10.3
python-dotenv==1.0.1
pandas==2.2.3
scikit-learn==1.5.2
openai>=1.0.0
spacy==3.7.5
